[Fix]: last try

+++ 0x1B-sorting_algorithms/107-quick_sort_hoare.c
@@ -8,10 +8,10 @@
  * @lo: first value
  * Return: integer.
  */
-int hoare_part(int *array, ssize_t lo, ssize_t hi, size_t size)
+int hoare_part(int *array, int lo, int hi, size_t size)
 {
 	int pivot;
-	ssize_t i, j;
+	int i, j;
 
 	i = lo - 1;
 	j = hi + 1;
@@ -45,9 +45,9 @@ int hoare_part(int *array, ssize_t lo, ssize_t hi, size_t size)
  * @lo: first value
  * Return: Nothing.
  */
-void quick_sort_easy_hoare(int *array, ssize_t lo, ssize_t hi, size_t size)
+void quick_sort_easy_hoare(int *array, int lo, int hi, size_t size)
 {
-	ssize_t pi;
+	int pi;
 
 	if (lo < hi)
 	{
@@ -65,8 +65,8 @@ void quick_sort_easy_hoare(int *array, ssize_t lo, ssize_t hi, size_t size)
  */
 void quick_sort_hoare(int *array, size_t size)
 {
-	ssize_t lo = 0;
-	ssize_t hi = size - 1;
+	int lo = 0;
+	int hi = size - 1;
 
 	if (array == NULL || size < 2)
 		return;
